+++ CHANGELOG.md
@@ -101,4 +101,5 @@
 * Temporary delete old `Persistable` instances
 
 # 0.2.3
-* Define `Persistable` instance for `Keyboard :&: Message` union
+* Define `Persistable` instance for `Send (Text :&: Keyboard)`
+* Move modified versions of `Capacity` and `Silently` datatypes to `Message` module